python wrapper for similarity transform

+++ wrapper/similarity_transform.cpp
@@ -1,8 +1,7 @@
 #include "similarity_transform.hpp"
 
 extern "C" int64_t max_eigen_value(float *mat, float *eigen_val,
-                                   float *eigen_vec, uint64_t dim,
-                                   uint *iter_cnt) {
+                                   float *eigen_vec, uint dim, uint *iter_cnt) {
   sycl::device d{sycl::default_selector{}};
   sycl::queue q{d};
